requirements.txt
psycopg2
yoyo-migrations
